+++ src/components/ChannelDetailsPanel.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+/**
+ * ChannelDetailsPanel component.
+ */
+const ChannelDetailsPanel = styled.section`
+  height: calc(100% - 30px);
+  overflow-y: auto;
+  padding: 10px;
+`
+
+export default ChannelDetailsPanel